refactor: rename `config.ts` to `site-config.ts` to match the YAML file
